CAAppKitBridge: In the demo, implement drawing a tree of CALayers owned by an NSView.
An NSView (temporarily) provides a CARenderer given it is the root of the tree, and its layer is being rendered in it in the demo.

This commit merges in the changes as of 049d520ebe1e7b7af9b059870cb2ab048bd80e9c on Tue, 31 Jul 2018 13:56:40 +0200, with added maintainer fixes from 45b85d149f86e9ae6e5f2b18d1d7b8b2676cb279 on Fri, 4 Apr 2025 18:41:59 +0100.

Fixes: #5
Co-authored-by: Ivan Vučica <ivan@vucica.net>